Fix pom resolution error for unrecognized tags
We've used setPomFile for the model building request when transitive
dependencies were resolved. This made the model builder operate in
project mode and caused some errors to be reported that complained about
unrecognized tags such as 'organization'.

These were not actual errors, however, we don't want the model builder
to operate in project mode so that needed to be turned off.