Fix for dependency resolution without <type> tag
When an artifact doesn't specify the <type> of its dependency, Maven
assumes jar by default. If it does that, then if the actual dependency
artifact has a different extension, then it won't be found.

As a workaround, the code will do adjustment for dependencies that have
jar type, and will modify its extension to the packaging type of the
resolved Model when appropriate.